separate utils function
